+++ cmd/secrets-provider/main.go
@@ -81,17 +81,15 @@ func main() {
 	}
 
 	// Gather secrets config and create a retryable Secrets Provider
-	provideSecrets, secretsConfig, err := retryableSecretsProvider(ctx, tracer, secretRetriever)
+	provideSecrets, _, err := retryableSecretsProvider(ctx, tracer, secretRetriever)
 	if err != nil {
 		logError(err.Error())
 		return
 	}
 
 	// Provide secrets
-	err = provideSecrets()
-	if err != nil {
-		errStr := fmt.Sprintf(messages.CSPFK039E, secretsConfig.StoreType, err.Error())
-		logError(errStr)
+	if err = provideSecrets(); err != nil {
+		logError(err.Error())
 	}
 }
 
@@ -129,11 +127,6 @@ func secretRetriever(ctx context.Context,
 	if err != nil {
 		span.RecordErrorAndSetStatus(err)
 		log.Error(messages.CSPFK008E)
-		return nil, err
-	}
-	if err = validateContainerMode(authnConfig.GetContainerMode()); err != nil {
-		span.RecordErrorAndSetStatus(err)
-		log.Error(err.Error())
 		return nil, err
 	}
 
@@ -187,6 +180,20 @@ func retryableSecretsProvider(
 		secretsConfig.RetryCountLimit,
 		provideSecrets,
 	)
+
+	// Create a channel to send a quit signal to the periodic secret provider.
+	// TODO: Currently, this is just used for testing, but in the future we
+	// may want to create a SIGTERM or SIGHUP handler to catch a signal from
+	// a user / external entity, and then send an (empty struct) quit signal
+	// on this channel to trigger a graceful shut down of the Secrets Provider.
+	providerQuit := make(chan struct{})
+
+	provideSecrets = secrets.PeriodicSecretProvider(
+		secretsConfig.SecretsRefreshInterval,
+		getContainerMode(),
+		provideSecrets,
+		providerQuit,
+	)
 	return provideSecrets, secretsConfig, nil
 }
 
@@ -233,16 +240,10 @@ func logErrorsAndInfos(errLogs []error, infoLogs []error) error {
 	return nil
 }
 
-func validateContainerMode(containerMode string) error {
-	validContainerModes := []string{
-		"init",
-		"application",
+func getContainerMode() string {
+	containerMode := "init"
+	if mode, exists := annotationsMap[secretsConfigProvider.ContainerModeKey]; exists {
+		containerMode = mode
 	}
-
-	for _, validContainerModeType := range validContainerModes {
-		if containerMode == validContainerModeType {
-			return nil
-		}
-	}
-	return fmt.Errorf(messages.CSPFK007E, containerMode, validContainerModes)
+	return containerMode
 }